refactor(exceptions): group exceptions in one file located in each package
