--- src/main/java/ch/sectioninformatique/auth/app/exceptions/InvalidCredentialsException.java
@@ -1,7 +1,0 @@
-package ch.sectioninformatique.auth.app.exceptions;
-
-public class InvalidCredentialsException extends AppException {
-    public InvalidCredentialsException() {
-        super("Invalid credentials");
-    }
-}
