--- src/main/java/ch/sectioninformatique/auth/app/exceptions/RoleNotFoundException.java
@@ -1,9 +1,0 @@
-package ch.sectioninformatique.auth.app.exceptions;
-
-import ch.sectioninformatique.auth.security.RoleEnum;
-
-public class RoleNotFoundException extends AppException {
-    public RoleNotFoundException(RoleEnum role) {
-        super("Role not found: " + role.name());
-    }
-}
