--- src/main/java/ch/sectioninformatique/auth/app/exceptions/SecurityException.java
@@ -1,8 +1,0 @@
-package ch.sectioninformatique.auth.app.exceptions;
-
-public class SecurityException extends AppException {
-    public SecurityException(String message) {
-        super(message);
-    }
-    
-}
