--- src/main/java/ch/sectioninformatique/auth/app/exceptions/UnauthorizedActionException.java
@@ -1,7 +1,0 @@
-package ch.sectioninformatique.auth.app.exceptions;
-
-public class UnauthorizedActionException extends AppException {
-    public UnauthorizedActionException(String message) {
-        super(message);
-    }
-}
