--- src/main/java/ch/sectioninformatique/auth/app/exceptions/UserAlreadyAdminException.java
@@ -1,8 +1,0 @@
-package ch.sectioninformatique.auth.app.exceptions;
-
-
-public class UserAlreadyAdminException extends AppException {
-    public UserAlreadyAdminException(String login) {
-        super("The user is already an admin: " + login);
-    }
-}
