--- src/main/java/ch/sectioninformatique/auth/app/exceptions/UserAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package ch.sectioninformatique.auth.app.exceptions;
-
-public class UserAlreadyExistsException extends AppException {
-    public UserAlreadyExistsException(String login) {
-        super("User already exists: " + login);
-    }
-}
