--- src/main/java/ch/sectioninformatique/auth/app/exceptions/UserAlreadyManagerException.java
@@ -1,7 +1,0 @@
-package ch.sectioninformatique.auth.app.exceptions;
-
-public class UserAlreadyManagerException extends AppException {
-    public UserAlreadyManagerException(String login) {
-        super("The user is already a manager: " + login);
-    }
-}
