--- src/main/java/ch/sectioninformatique/auth/app/exceptions/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package ch.sectioninformatique.auth.app.exceptions;
-
-public class UserNotFoundException extends AppException {
-    public UserNotFoundException(String loginOrId) {
-        super("User not found: " + loginOrId);
-    }
-}
